fix notify to not block clicks
The notify toast is always on screen, also when hidden, and blocks clicks on the bottom right of the screen with its padding. After it's hidden, the content stays in it and blocks an even larger area.

So we can't check for `:empty`, we always exclude it from mouse clicks instead.