feat(mobile) style feeds toggle as floating action button
This provides maximum usable width for phones. Tablets and bigger screens still have the bar on the left side, should be more ergonomic there since feeds open on the left side and users' hands are usually further up. This also prevents size calculation bugs when rotating the device. Phones are hardly usable in landscape mode anyway because of too little height in that orientation.

This also aligns the action button with article action icons on all touch screens and updates the feeds toggle icon.